Implement basic arithmetic and bitwise operators (#295)
Implement translations for the basic arithmetic and bitwise operators
such as +, - * et al.

Also implements explicit precedence via parentheses, which needs some
attention in the grammar to be done perfectly, but that is for another
PR.

As part of generally processing constant expressions, a NoTable AST
construct is added as not all statements require a FROM clause, the
SELECT being most pertinent here.

we also clean up some of the visitor code such that unnecessary
overrides are removed. The ANTLR4 visitor already calls accept()
automatically so there is no need to implement an override to do exactly
the same.

---------

Signed-off-by: Jim.Idle <jimi@idle.ws>