feat: Implement freetext function visitor, simplify grammar (#487)
Freetext functions, $PARTITION and HIERARCHYID functions are unsupported
in Databricks SQL, but we must still support generation of IR to
indicate this. We also therefore prevent errors when visiting
functionCall.

Here we complete handling of all TSQL functions from a parsing and IR
generating point of view. There are likely to be many conversions of
functions and flagging of unsupported functions as the coverage tests
are expanded to support them.

Progresses: #275 
Closes: #357