Greatly improve grammar for CREATE/ALTER (#805)
Here, we greatly improve the grammar constructs for create and alter and
additionally, make parsing of `id . id` orthogonal instead of having 17
rules to parse the same construct.

We also remove the mountain of superfluous labels given to many tokens
and rules, which just add bloat to the generated code.

As a result, the error recovery mechanism is more accurate in
identifying the source of syntax errors as there is much less ambiguity.
The size of the generated parser file is also 'smaller' at 99387 lines
vs 107354 - we intend to continue reducing this.