Create repeatable estimator for Snowflake query history (#924)
_*NB*: This branch is based upon:
[chore/coverage-in-core](https://github.com/databrickslabs/remorph/tree/chore/coverage-in-core)
and will appear to contain all those changes until that branch is merged
into main and this branch is rebased)._

Here we implement a coverage estimation tool allowing an analysis of the
query history of a given database. Given connection details for a
system, and its dialect (only Snowflake is implemented here as a first
step), it will report how much of the query history it is able to parse
and transpile into Databricks SQL.

While the estimator is currently implemented only for Snowflake, it is
intended that a similar report will be provided for each supported
database/dialect.

A subsequent enhancement will build on this PR to report on SQL
complexity.

---------

Co-authored-by: Valentin Kasas <valentin.kasas@xebia.com>