Return `Result` out of generators, instead of throwing exceptions (#1006)
This is a rather important change although the most part (changes in the
`Generator`s) is quite mechanical.

The signatures in the `Generator` trait have been changed so that the
`generate` method returns a `Result[Out]` (instead of just `Out`
before).

This introduces a difficulty, as the result of recursive calls can no
longer be used directly, but must be accounted for potential failures by
the caller. To make it easier to work with `Result[String]` (which have
been aliased to `SQL`), an `SQLInterpolater` has been added (see
`SQLInterpolatorSpec` for examples of how it can be used), as well as
`mkSql` extension methods for `Seq[SQL]` (similar to the way `mkString`
works on `Seq[String]`).

Finally, we change `ReportEntry` produced by coverage tests to contain
`RemorphError`s instead of JSON strings, to open the way for more
structured and more informative such reports (which was the initial
trigger for this whole change).