Rework error return system and use it to generate parsing error and coverage stats (#1004)
Here we stop converting error messages from the parse, IR builders and
source generators to strings and JSON strings prematurely. Errors are
now returned as a `RemorphError` and then that is converted to JSON
where it is appropriate.

The estimation system now preserves the error structure, including the
ruleName and `TOKEN` when a parsing error occurs and similarly for
errors produced when building IR and finding missing visitors etc.