Introduce an abstraction for handling stateful traversal of the tree (#1018)
While still being able to accumulate errors, produce partial results,
etc.

The base principle is to build a *representation* of the computation
(parsing, visiting, optimizing, generating, etc) that will eventually
get `run`. This representation keeps track of a state (of arbitrary type
if we wanted to, but we practically stick to `RemorphContext` for now)
along with the result it is producing.

This way, we can interleave any step of the computation with updates to
this state, like counting the statements processed thus far (for any
definition of what a statement is, including arbitrarily nested
subqueries), the branch of the IR tree currently being processed, etc.

Similarly, at any point of the computation, we can inspect the current
state, to guide said computation (by customizing error messages with
contextual information for example).