Refactor `x.tail.foldleft(x.head)` chains as `x.reduceLeft` (#1249)
This PR refactors the situations where we use this:

```scala
x.tail.foldLeft(x.head)
```

to use the equivalent operation that does this:
```scala
x.reduceLeft
```

Following a review comment discussing the safety of reduce in these
situations[^1], some further changes have been made to the surrounding
code such that the behaviour remains the same but the safety can be
determined via trivial inspection without needing to refer to the
grammar.

[^1]: Not a new concern; the previous code had the same safety issue due
to the use of `.head`.