Create preprocessor and workflow for Jinja template processing (#1168)
A preprocessor and workflow to allow conversion of embedded SQL in Jinja
templates while preserving the templates and, as much as possible, the
original text format.

See the preprocessors/README.md for full details of the implementation.

Note that this PR implements about 95% of the use cases for templates,
but there are likely other places to support them, such as `if`
templates in expression lists. But we will now address these on a case
by case basis throuhg testing.

NOTE: This PR only supports templates in TSQL, but it is already too
big, so Snowflake support will be added in a separate PR, which is now
trivial to do, but would add another 70+ lines of code changes to this
PR.

---------

Co-authored-by: Valentin Kasas <valentin.kasas@xebia.com>