Extend recon queries to use normalized column names (#1923)
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary, they're helpful to illustrate the before and
after state -->
### What does this PR do?
**Big Picture:** 
Add support in lakebridge reconciler for columns with special chars e.g.
`Cust#` as at the moment reconciler will fail if such columns are
passed.

**This PR:**
The PR adds support for delimiting all identifiers in the SQL queries
produced by the different `QueryBuilder`s
This is necessary in case a column has a special character in the name

### Relevant implementation details
1. We use spark sql on top of JDBC to execute those queries. This means
that the initial reading of the data from the databases is deferred to
the source databases
2. SQL is sent directly to the database in that database’s own dialect
(e.g., use [Name] for SQL Server identifiers).
3. Spark wraps the provided SQL in a subquery so it can treat results
like a table, but the inner SQL is executed by the database.
4. Inside Spark, queries are written in Spark’s ANSI-like dialect. e.g.
compare hash queries


### Caveats/things to watch out for when reviewing:
This is a big change that had the implementation split between 6 PRs to
make reviews manageable. this cannot go to main without the rest of the
PRs. this is PR 6/6

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #1866 

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs lakebridge ...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [ ] manually tested
- [x] added unit tests
- [x] added integration tests