Fix conversion output when a nested directory structure is present. (#2131)
## Changes

This PR fixes a bug whereby files nested 2 or more directories deep
within the input directory could fail to be written out after
conversion. This could occur when the directory structure wasn't already
in place. (Sometimes conversion of multiple files at a higher level
might create the directories, but this was accidental and can't be
relied upon.)

### Relevant implementation details

The fix is trivial; testing it is more work than fixing it.

### Linked issues

Resolves #1945

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- added unit test