Hotfix: increase the maximum line-size that can be handled when process stderr from LSP servers (#2160)
## Changes

### What does this PR do?

During `transpile` anything written by the LSP server to stderr is
mirrored in the logs. Due to the internal implementation, there's a
default maximum size of 64K for each line. Recent versions of BB can
output lines that are larger than this limit, which causes a crash or a
hang. This PR is a hot fix to increase the limit, pending tests and a
proper fix.

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- manually tested