Improve handling of output from LSP servers (#2163)
## Changes

LSP servers use stdin/stdout for the LSP calls, and stderr as a general
logging facility. Currently we mirror the stderr output as logs, but due
to the implementation details this fails when an individual line
received from stderr is longer than 64KiB in size.

Although #2160 contains a hot fix for this problem, this PR improves the
handling in a more robust manner:

- Long lines from the LSP server are now detected and broken up into
chunks, as several log entries. (This is preferable to a large limit
because it means we have an upper bound on memory usage rather than
potentially buffering forever and running out of memory.)
- The cause of the error when a long line hits has been addressed,
however if we hit some other error during stderr processing we now log
the (critical!) error so that it's at least clear something has gone
wrong. (The nature of this is that the CLI might hang, but it will be
noisy about it.)

### Relevant implementation details

The primary change here is to handle line-breaking and decoding
ourselves rather than relying on the `StreamReader.readline()`
implementation. (The behaviour of the latter cannot be reliably
controlled when a line exceeds the configured `limit`, which defaults to
64K.)

### Caveats/things to watch out for when reviewing:

Further changes are needed to the test LSP server (`lsp_server.py`) that
we use for testing, they are out of scope for this PR.

### Linked issues

Supersedes #2160, closes #2164.

This PR is the basis for databrickslabs/blueprint#320: once that's
available the changes in here can be simplified/refactored.

### Functionality

- modified existing command: `databricks labs lakebridge transpile`

### Tests

- manually tested
- added unit tests