Improve launch of python-based LSP server within IDEs (#2184)
## Changes

This PR tweaks the way we launch python-based LSP launchers, mainly to
help with developer workflows in the latest IntelliJ release. The issue
here is that when running individual tests via the IDE, as of the latest
release the (hatch-created) virtual environment is not activated: this
causes many tests to run the stub LSP server against the system python,
or fail because the system only provides `python3` (and not `python`).

### Relevant implementation details

The key change here is that if the LSP server is python-based but
there's no dedicated `.venv` for it, the LSP server is now always
launched with the current python interpreter.

As a related minor improvement, we also raise an error earlier if the
executable we're about to launch can't be found.

### Tests

- manually tested
- existing unit tests
- existing integration tests