Implement subset of `workspace/applyEdit` for future conversion output (#2201)
## Changes

### What does this PR do?

This PR introduces a basic implementation for LSP's
`workspace/applyEdit`, supporting the subset we need for replacing a
file.

This editor is not wired up to the LSP server, that will be implemented
in the next PR. The scope is specifically limited to the document
changes needed to replace a file in its entirely by issuing sequences
of:

 - `CreateFile`: create/truncate a file, preparing to write the content.
 - `TextDocumentEdit`: writing the content into the file.

The editor enforces:

 - UTF-8 encoding of written files.
 - Line normalisation: all lines are `\n`-denoted.

This is the first step to allowing LSP servers to use
`workspace/applyEdit` to write out converted files rather than returning
them as part of the current custom protocol.

### Relevant implementation details

The implementation goes to a lot of lengths to ensure that:

 - We specifically support this sequence of operations.
- Unsupported parts of the LSP protocol are explicitly handled as
failures.
 - Unhappy (I/O-related) paths are handled.

### Tests

- added unit tests

---------

Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>